+++ src/lib/codegenerator/generator.test.js
@@ -1536,6 +1536,7 @@ test('Generate code for a template with custom components', (assert) => {
     elementConfigs[1] = {}
 
     elms[1] = this.element({ parent: parent || 'root' }, inSlot === true ? slotComponent : component)
+    elementConfigs[1]['holder'] = true
     skips[1] = []
     if (typeof cmps[1] !== 'undefined') {
         for (let key in cmps[1][Symbol.for('config')].props) {
@@ -1580,6 +1581,7 @@ test('Generate code for a template with custom components', (assert) => {
     elementConfigs[3] = {}
 
     elms[3] = this.element({ parent: parent || 'root' }, inSlot === true ? slotComponent : component)
+    elementConfigs[3]['holder'] = true
     skips[3] = []
     if (typeof cmps[3] !== 'undefined') {
         for (let key in cmps[3][Symbol.for('config')].props) {
@@ -1716,6 +1718,7 @@ test('Generate code for a template with an unregistered custom component', (asse
     parent = elms[0]
     elementConfigs[1] = {}
     elms[1] = this.element({ parent: parent || 'root' }, inSlot === true ? slotComponent : component)
+    elementConfigs[1]['holder'] = true
     skips[1] = []
     if (typeof cmps[1] !== 'undefined') {
         for (let key in cmps[1][Symbol.for('config')].props) {
@@ -1756,6 +1759,7 @@ test('Generate code for a template with an unregistered custom component', (asse
     parent = elms[0]
     elementConfigs[3] = {}
     elms[3] = this.element({ parent: parent || 'root' }, inSlot === true ? slotComponent : component)
+    elementConfigs[3]['holder'] = true
     skips[3] = []
     if (typeof cmps[3] !== 'undefined') {
         for (let key in cmps[3][Symbol.for('config')].props) {
@@ -1889,6 +1893,7 @@ test('Generate code for a template with custom components with arguments', (asse
     parent = elms[0]
     elementConfigs[1] = {}
     elms[1] = this.element({ parent: parent || 'root' }, inSlot === true ? slotComponent : component)
+    elementConfigs[1]['holder'] = true
     elementConfigs[1]['x'] = 10
     skips[1] = []
     if (typeof cmps[1] !== 'undefined') {
@@ -1931,6 +1936,7 @@ test('Generate code for a template with custom components with arguments', (asse
     parent = elms[0]
     elementConfigs[3] = {}
     elms[3] = this.element({ parent: parent || 'root' }, inSlot === true ? slotComponent : component)
+    elementConfigs[3]['holder'] = true
     elementConfigs[3]['x'] = 100
     elementConfigs[3]['img'] = component.img
     skips[3] = []
@@ -2071,6 +2077,7 @@ test('Generate code for a template with custom components with argument value as
 
     elms[1] = this.element({ parent: parent || 'root' }, inSlot === true ? slotComponent : component)
 
+    elementConfigs[1]['holder'] = true
     elementConfigs[1]['x'] = 1000
     elementConfigs[1]['y'] = 100
     elementConfigs[1]['h'] = parent.node.height * (45 / 100)
@@ -2223,6 +2230,7 @@ test('Generate code for a template with custom components with reactive props', 
     parent = elms[0]
     elementConfigs[1] = {}
     elms[1] = this.element({ parent: parent || 'root' }, inSlot === true ? slotComponent : component)
+    elementConfigs[1]['holder'] = true
     elementConfigs[1]['x'] = 10
     elementConfigs[1]['img'] = component.image
     skips[1] = []
@@ -2272,6 +2280,7 @@ test('Generate code for a template with custom components with reactive props', 
     parent = elms[0]
     elementConfigs[3] = {}
     elms[3] = this.element({ parent: parent || 'root' }, inSlot === true ? slotComponent : component)
+    elementConfigs[3]['holder'] = true
     elementConfigs[3]['x'] = 100
     elementConfigs[3]['img'] = component.image
     skips[3] = []
@@ -2538,6 +2547,7 @@ test('Generate code for a template with slot content', (assert) => {
     parent = elms[0]
     elementConfigs[1] = {}
     elms[1] = this.element({ parent: parent || 'root' }, inSlot === true ? slotComponent : component)
+    elementConfigs[1]['holder'] = true
     elementConfigs[1]['w'] = 1920
     elementConfigs[1]['h'] = 1080
     skips[1] = []
@@ -2708,6 +2718,7 @@ test('Generate code for a template with slot content, using a named slot', (asse
     parent = elms[0]
     elementConfigs[1] = {}
     elms[1] = this.element({ parent: parent || 'root' }, inSlot === true ? slotComponent : component)
+    elementConfigs[1]['holder'] = true
     elementConfigs[1]['w'] = 1920
     elementConfigs[1]['h'] = 1080
     skips[1] = []
@@ -4264,6 +4275,7 @@ test('Generate code for a template with a simple for-loop on a Component with a 
           )
         }
 
+        elementConfigs[1]['holder'] = true
         skips[1] = []
         if (typeof cmps[1] !== 'undefined') {
           for (let key in cmps[1][Symbol.for('config')].props) {